Extract MapContainsKey validation to common lib
