+++ cab-app-common/src/main/resources/ValidationMessages.properties
@@ -1,0 +1,1 @@
+by.arvisit.cabapp.common.validation.MapContainsKey.message=Patch should contain key: {value}
