+++ cab-app-driver-service/src/main/resources/messages.properties
@@ -26,7 +26,4 @@ by.arvisit.cabapp.driverservice.controller.DriverController.patch.NotNull.messag
 by.arvisit.cabapp.driverservice.controller.DriverController.patch.Size.message=Patch should contain one key:value pair
 by.arvisit.cabapp.driverservice.controller.DriverController.patch.isAvailable.NotNull.message=isAvailable value should not be null
 
-
-by.arvisit.cabapp.driverservice.validation.MapContainsKey.message=Patch should contain key: {value}
-
 by.arvisit.cabapp.common.dto.ListContainerResponseDto.DriverResponseDto.IllegalStateException.template=No available drivers right now
