+++ cab-app-rides-service/src/main/resources/messages.properties
@@ -38,5 +38,3 @@ by.arvisit.cabapp.ridesservice.persistence.model.Ride.passengerScore.IllegalStat
 by.arvisit.cabapp.ridesservice.persistence.model.Ride.driverScore.EntityNotFoundException.template=Found no scores for driver with id {0}
 by.arvisit.cabapp.ridesservice.persistence.model.Ride.passengerScore.EntityNotFoundException.template=Found no scores for passenger with id {0}
 by.arvisit.cabapp.ridesservice.persistence.model.Ride.isPaid.IllegalStateException.template=Payment has not gotten SUCCESS status for some reason
-
-by.arvisit.cabapp.ridesservice.validation.MapContainsKey.message=Patch should contain key: {value}
